Move code in semanticdb/scalac to scala.meta.internal.semanticdb.scalac
