+++ semanticdb/scalac/library/src/main/scala/scala/meta/internal/semanticdb/scalac/HijackAnalyzer.scala
@@ -1,4 +1,4 @@
-package scala.meta.internal
+package scala.meta.internal.semanticdb.scalac
 
 import scala.collection.mutable
 import scala.tools.nsc.{Global => NscGlobal, SubComponent}
