+++ semanticdb/scalac/library/src/main/scala/scala/meta/internal/semanticdb/scalac/LengthWriter.scala
@@ -1,4 +1,4 @@
-package scala.meta.internal.semanticdb
+package scala.meta.internal.semanticdb.scalac
 
 import java.io.Writer
