+++ semanticdb/scalac/library/src/main/scala/scala/meta/internal/semanticdb/scalac/ParseOps.scala
@@ -1,5 +1,4 @@
-package scala.meta.internal
-package semanticdb
+package scala.meta.internal.semanticdb.scalac
 
 import scala.{meta => m}
