+++ semanticdb/scalac/library/src/main/scala/scala/meta/internal/semanticdb/scalac/PrinterOps.scala
@@ -1,4 +1,4 @@
-package scala.meta.internal.semanticdb
+package scala.meta.internal.semanticdb.scalac
 
 import scala.{meta => m}
 import java.io.PrintWriter
